requirements.txt
# requirements.txt
requests
python-telegram-bot